Move files in attributes_updater dir to model_node dir.
