+++ src/main/cc/wfa/virtual_people/core/model/attributes_updater.cc
@@ -12,17 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "wfa/virtual_people/core/model/attributes_updater/attributes_updater.h"
+#include "wfa/virtual_people/core/model/attributes_updater.h"
 
 #include "absl/container/flat_hash_map.h"
 #include "absl/status/status.h"
 #include "absl/status/statusor.h"
 #include "src/main/proto/wfa/virtual_people/common/model.pb.h"
-#include "wfa/virtual_people/core/model/attributes_updater/conditional_assignment_impl.h"
-#include "wfa/virtual_people/core/model/attributes_updater/conditional_merge_impl.h"
-#include "wfa/virtual_people/core/model/attributes_updater/sparse_update_matrix_impl.h"
-#include "wfa/virtual_people/core/model/attributes_updater/update_matrix_impl.h"
-#include "wfa/virtual_people/core/model/attributes_updater/update_tree_impl.h"
+#include "wfa/virtual_people/core/model/conditional_assignment_impl.h"
+#include "wfa/virtual_people/core/model/conditional_merge_impl.h"
+#include "wfa/virtual_people/core/model/sparse_update_matrix_impl.h"
+#include "wfa/virtual_people/core/model/update_matrix_impl.h"
+#include "wfa/virtual_people/core/model/update_tree_impl.h"
 #include "wfa/virtual_people/core/model/model_node.h"
 
 namespace wfa_virtual_people {
