--- src/main/cc/wfa/virtual_people/core/model/attributes_updater/BUILD.bazel
@@ -1,97 +1,0 @@
-load("@rules_cc//cc:defs.bzl", "cc_test")
-
-package(default_visibility = ["//src:__subpackages__"])
-
-exports_files([
-    "attributes_updater.cc",
-    "attributes_updater.h",
-    "conditional_assignment_impl.cc",
-    "conditional_assignment_impl.h",
-    "conditional_merge_impl.cc",
-    "conditional_merge_impl.h",
-    "sparse_update_matrix_impl.cc",
-    "sparse_update_matrix_impl.h",
-    "update_matrix_impl.cc",
-    "update_matrix_impl.h",
-    "update_tree_impl.cc",
-    "update_tree_impl.h",
-])
-
-cc_test(
-    name = "update_matrix_impl_test",
-    srcs = ["update_matrix_impl_test.cc"],
-    deps = [
-        "//src/main/cc/wfa/virtual_people/core/model:model_node",
-        "@com_google_absl//absl/container:flat_hash_map",
-        "@com_google_absl//absl/status:status",
-        "@com_google_absl//absl/status:statusor",
-        "@com_google_googletest//:gtest_main",
-        "@com_google_protobuf//:protobuf",
-        "@cross_media_measurement//src/test/cc/testutil:matchers",
-        "@cross_media_measurement//src/test/cc/testutil:status_macros",
-        "@virtual_people_common//src/main/proto/wfa/virtual_people/common:model_cc_proto",
-    ],
-)
-
-cc_test(
-    name = "update_tree_impl_test",
-    srcs = ["update_tree_impl_test.cc"],
-    deps = [
-        "//src/main/cc/wfa/virtual_people/core/model:model_node",
-        "@com_google_absl//absl/container:flat_hash_map",
-        "@com_google_absl//absl/status:status",
-        "@com_google_absl//absl/status:statusor",
-        "@com_google_googletest//:gtest_main",
-        "@com_google_protobuf//:protobuf",
-        "@cross_media_measurement//src/test/cc/testutil:matchers",
-        "@cross_media_measurement//src/test/cc/testutil:status_macros",
-        "@virtual_people_common//src/main/proto/wfa/virtual_people/common:model_cc_proto",
-    ],
-)
-
-cc_test(
-    name = "sparse_update_matrix_impl_test",
-    srcs = ["sparse_update_matrix_impl_test.cc"],
-    deps = [
-        "//src/main/cc/wfa/virtual_people/core/model:model_node",
-        "@com_google_absl//absl/container:flat_hash_map",
-        "@com_google_absl//absl/status:status",
-        "@com_google_absl//absl/status:statusor",
-        "@com_google_googletest//:gtest_main",
-        "@com_google_protobuf//:protobuf",
-        "@cross_media_measurement//src/test/cc/testutil:matchers",
-        "@cross_media_measurement//src/test/cc/testutil:status_macros",
-        "@virtual_people_common//src/main/proto/wfa/virtual_people/common:model_cc_proto",
-    ],
-)
-
-cc_test(
-    name = "conditional_merge_impl_test",
-    srcs = ["conditional_merge_impl_test.cc"],
-    deps = [
-        "//src/main/cc/wfa/virtual_people/core/model:model_node",
-        "@com_google_absl//absl/status:status",
-        "@com_google_absl//absl/status:statusor",
-        "@com_google_googletest//:gtest_main",
-        "@com_google_protobuf//:protobuf",
-        "@cross_media_measurement//src/test/cc/testutil:matchers",
-        "@cross_media_measurement//src/test/cc/testutil:status_macros",
-        "@virtual_people_common//src/main/proto/wfa/virtual_people/common:model_cc_proto",
-    ],
-)
-
-cc_test(
-    name = "conditional_assignment_impl_test",
-    srcs = ["conditional_assignment_impl_test.cc"],
-    deps = [
-        "//src/main/cc/wfa/virtual_people/core/model:model_node",
-        "@com_google_absl//absl/status:status",
-        "@com_google_absl//absl/status:statusor",
-        "@com_google_googletest//:gtest_main",
-        "@com_google_protobuf//:protobuf",
-        "@cross_media_measurement//src/test/cc/testutil:matchers",
-        "@cross_media_measurement//src/test/cc/testutil:status_macros",
-        "@virtual_people_common//src/main/proto/wfa/virtual_people/common:demographic_cc_proto",
-        "@virtual_people_common//src/main/proto/wfa/virtual_people/common:model_cc_proto",
-    ],
-)
