+++ src/main/cc/wfa/virtual_people/core/model/branch_node_impl.h
@@ -21,7 +21,7 @@
 #include "absl/strings/string_view.h"
 #include "absl/types/variant.h"
 #include "src/main/proto/wfa/virtual_people/common/model.pb.h"
-#include "wfa/virtual_people/core/model/attributes_updater/attributes_updater.h"
+#include "wfa/virtual_people/core/model/attributes_updater.h"
 #include "wfa/virtual_people/core/model/model_node.h"
 #include "wfa/virtual_people/core/model/utils/distributed_consistent_hashing.h"
 #include "wfa/virtual_people/core/model/utils/field_filters_matcher.h"
