+++ src/main/cc/wfa/virtual_people/core/model/conditional_assignment_impl.h
@@ -12,15 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef WFA_VIRTUAL_PEOPLE_CORE_MODEL_ATTRIBUTES_UPDATER_CONDITIONAL_ASSIGNMENT_IMPL_H_
-#define WFA_VIRTUAL_PEOPLE_CORE_MODEL_ATTRIBUTES_UPDATER_CONDITIONAL_ASSIGNMENT_IMPL_H_
+#ifndef WFA_VIRTUAL_PEOPLE_CORE_MODEL_CONDITIONAL_ASSIGNMENT_IMPL_H_
+#define WFA_VIRTUAL_PEOPLE_CORE_MODEL_CONDITIONAL_ASSIGNMENT_IMPL_H_
 
 #include "absl/status/status.h"
 #include "absl/status/statusor.h"
 #include "google/protobuf/descriptor.h"
 #include "src/main/proto/wfa/virtual_people/common/model.pb.h"
 #include "wfa/virtual_people/common/field_filter/field_filter.h"
-#include "wfa/virtual_people/core/model/attributes_updater/attributes_updater.h"
+#include "wfa/virtual_people/core/model/attributes_updater.h"
 
 namespace wfa_virtual_people {
 
@@ -75,4 +75,4 @@ class ConditionalAssignmentImpl : public AttributesUpdaterInterface {
 
 }  // namespace wfa_virtual_people
 
-#endif  // WFA_VIRTUAL_PEOPLE_CORE_MODEL_ATTRIBUTES_UPDATER_CONDITIONAL_ASSIGNMENT_IMPL_H_
+#endif  // WFA_VIRTUAL_PEOPLE_CORE_MODEL_CONDITIONAL_ASSIGNMENT_IMPL_H_
