+++ src/main/cc/wfa/virtual_people/core/model/conditional_merge_impl.h
@@ -12,13 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef WFA_VIRTUAL_PEOPLE_CORE_MODEL_ATTRIBUTES_UPDATER_CONDITIONAL_MERGE_IMPL_H_
-#define WFA_VIRTUAL_PEOPLE_CORE_MODEL_ATTRIBUTES_UPDATER_CONDITIONAL_MERGE_IMPL_H_
+#ifndef WFA_VIRTUAL_PEOPLE_CORE_MODEL_CONDITIONAL_MERGE_IMPL_H_
+#define WFA_VIRTUAL_PEOPLE_CORE_MODEL_CONDITIONAL_MERGE_IMPL_H_
 
 #include "absl/status/status.h"
 #include "absl/status/statusor.h"
 #include "src/main/proto/wfa/virtual_people/common/model.pb.h"
-#include "wfa/virtual_people/core/model/attributes_updater/attributes_updater.h"
+#include "wfa/virtual_people/core/model/attributes_updater.h"
 #include "wfa/virtual_people/core/model/utils/field_filters_matcher.h"
 
 namespace wfa_virtual_people {
@@ -71,4 +71,4 @@ class ConditionalMergeImpl : public AttributesUpdaterInterface {
 
 }  // namespace wfa_virtual_people
 
-#endif  // WFA_VIRTUAL_PEOPLE_CORE_MODEL_ATTRIBUTES_UPDATER_CONDITIONAL_MERGE_IMPL_H_
+#endif  // WFA_VIRTUAL_PEOPLE_CORE_MODEL_CONDITIONAL_MERGE_IMPL_H_
