+++ src/main/cc/wfa/virtual_people/core/model/update_tree_impl.h
@@ -12,14 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef WFA_VIRTUAL_PEOPLE_CORE_MODEL_ATTRIBUTES_UPDATER_UPDATE_TREE_IMPL_H_
-#define WFA_VIRTUAL_PEOPLE_CORE_MODEL_ATTRIBUTES_UPDATER_UPDATE_TREE_IMPL_H_
+#ifndef WFA_VIRTUAL_PEOPLE_CORE_MODEL_UPDATE_TREE_IMPL_H_
+#define WFA_VIRTUAL_PEOPLE_CORE_MODEL_UPDATE_TREE_IMPL_H_
 
 #include "absl/container/flat_hash_map.h"
 #include "absl/status/status.h"
 #include "absl/status/statusor.h"
 #include "src/main/proto/wfa/virtual_people/common/model.pb.h"
-#include "wfa/virtual_people/core/model/attributes_updater/attributes_updater.h"
+#include "wfa/virtual_people/core/model/attributes_updater.h"
 #include "wfa/virtual_people/core/model/model_node.h"
 
 namespace wfa_virtual_people {
@@ -52,4 +52,4 @@ class UpdateTreeImpl : public AttributesUpdaterInterface {
 
 }  // namespace wfa_virtual_people
 
-#endif  // WFA_VIRTUAL_PEOPLE_CORE_MODEL_ATTRIBUTES_UPDATER_UPDATE_TREE_IMPL_H_
+#endif  // WFA_VIRTUAL_PEOPLE_CORE_MODEL_UPDATE_TREE_IMPL_H_
